Add files via upload
This is a restaurant management system for reservations in C# using a .NET framework. It includes a login and registration page for administrators and staff. The admin can only edit (add, delete, remove) users from the system. The database is well organized, and have been using triggers for some tables too. Hope you enjoy it!